🚚  F06: Moved <Navbar/> and <UserConnected/> to the Navbar folder and updated the relative paths
