+++ src/components/hoc/Navbar/UserConnected.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import { theme } from "../../theme/theme";
 import { Link } from "react-router-dom";
+import { theme } from "../../../theme/theme";
 import { BsPersonCircle } from "react-icons/bs";
 
 export default function UserConnected({ userName }) {
